+++ data_preprocessing/seoul_basic_preprocessing.py
@@ -1,5 +1,6 @@
 import pandas as pd
-from datetime import datetime
+import re
+from datetime import datetime, timedelta
 
 
 '''
@@ -31,23 +32,83 @@ output_csv_file = 'preprocessing_csv_files/seoul_base_preprocessing.csv'
 seoul_api_data = pd.read_csv(input_csv_file)
 
 
-# 출생연도 컬럼의 날짜 데이터를 지정된 형식으로 변환
-# seoul_api_data 데이터프레임의 '태어난년도' 열에 대해 반복문 적용
+# # ID 컬럼에서 영어 문자가 포함된 행 삭제 : 해당 값을 확인 했을 때 오류 행으로 확인됨
+seoul_api_data = seoul_api_data[~seoul_api_data['ID'].apply(lambda x: bool(re.search(r'[a-zA-Z]', str(x))))]
+
+
+# '종' 컬럼에서 값이 'DOG'인 행만 유지 : 주제가 '유기견'이므로 기타 다른 동물 데이터에 대한 필터링
+seoul_api_data = seoul_api_data[seoul_api_data['종'] == 'DOG'].reset_index(drop=True)
+
+
+# # '종' 컬럼의 'DOG' 값을 '믹스'로 변경 : 논의 필요, 알 수 없는 견종에서 믹스가 가장 많아 해당 전처리를 수행했던 것으로 추정
+seoul_api_data['종'] = seoul_api_data['종'].replace('DOG', '믹스')
+
+
+# '출생연도' 컬럼 생성
+entry_dates = seoul_api_data['들어온 날짜']
+ages = seoul_api_data['나이']
+birth_years = []
+
 for i in range(len(seoul_api_data)):
-    date_str = seoul_api_data['태어난년도'][i]
+    try:
+        # 들어온 날짜를 datetime 형식으로 변환
+        entry_date = datetime.strptime(entry_dates[i], '%Y-%m-%d')
 
-    # 문자열을 datetime 객체로 변환
-    date_obj = datetime.strptime(date_str, '%Y/%m/%d %H:%M')
+        # 나이 데이터에서 "세"와 "개월"을 추출
+        age_info = ages[i]
+        if '(' in age_info and ')' in age_info:
+            # "2(세) 2(개월)" 형태에서 숫자를 추출
+            try:
+                age_years = int(age_info.split('(세)')[0].strip())
+                age_months = int(age_info.split('(세)')[1].replace('(개월)', '').strip())
+            except ValueError:
+                # 숫자로 변환하지 못하면 'Unknown'으로 처리
+                birth_years.append('Unknown')
+                continue
 
-    # 새로운 형식의 문자열로 변환 (e.g. '2023.11')
-    formatted_date = date_obj.strftime('%Y.%m')
+            # 출생 날짜 계산
+            birth_date = entry_date - timedelta(days=(age_years * 365 + age_months * 30.4))
+            birth_year = birth_date.year
+            birth_month = birth_date.month
 
-    # 변환된 값으로 컬럼 값 업데이트
-    seoul_api_data['태어난년도'][i] = formatted_date
+            # 연도와 월 형식을 보정
+            if birth_year < 100:  # 두 자리 연도일 경우
+                birth_year = 2000 + birth_year
+
+            birth_years.append(f"{birth_year}.{birth_month:02d}")  # YYYY.MM 형식
+        else:
+            # 나이 데이터 형식이 맞지 않을 경우 'Unknown' 처리
+            birth_years.append('Unknown')
+
+    except ValueError as e:
+        # 들어온 날짜 형식이 잘못된 경우 예외 처리
+        print(f"Date parsing error at row {i}: {e}")
+        birth_years.append('Unknown')
+
+# '출생연도' 컬럼 추가
+seoul_api_data['출생연도'] = birth_years
+
+
+# 보호장소 컬럼 추가
+locations = []
+for name in seoul_api_data['이름']:
+    if '(' in name and '센터' in name:
+        start_index = name.find('(') + 1  # 괄호 시작 위치
+        end_index = name.find('센터')  # "센터" 이전 위치
+        location = name[start_index:end_index] + '구'  # 추출된 지역명에 "구" 추가
+        locations.append(location)
+    else:
+        locations.append('Unknown')  # 형식이 맞지 않는 경우
+
+seoul_api_data['보호장소'] = locations
 
 
 # '보호 장소' 열의 값을 '서울특별시' + {x}의 형태로 통일 (e.g. 마포구 -> 서울특별시 마포구)
 seoul_api_data['보호장소'] = seoul_api_data['보호장소'].apply(lambda x: '서울특별시' if '서울시' in str(x) else '서울특별시 ' + str(x))
+
+
+# '이름' 컬럼에서 ()와 그 안의 값 제거
+seoul_api_data['이름'] = seoul_api_data['이름'].apply(lambda x: re.sub(r'\(.*?\)', '', x).strip())
 
 
 # 새로운 '세부견종' 열 생성 및 데이터 분리
@@ -60,28 +121,36 @@ for i in range(len(seoul_api_data)):
 
     # 괄호가 있는 경우 세부 견종 분리
     if index_bracket != -1:
-        sub_breeds = sample[index_bracket + 1:-2].replace('믹스','')
+        sub_breeds = sample[index_bracket + 1:-2].replace('믹스', '')
         if '+' in sub_breeds:
             sub_breeds_list = sub_breeds.split('+')
-            seoul_api_data['종'][i] = '믹스'
-            seoul_api_data['세부견종'][i] = ', '.join(sub_breeds_list)
+            seoul_api_data.loc[i, '종'] = '믹스'
+            seoul_api_data.loc[i, '세부견종'] = ', '.join(sub_breeds_list)
         elif '*' in sub_breeds:
             sub_breeds_list = sub_breeds.split('*')
-            seoul_api_data['종'][i] = '믹스'
-            seoul_api_data['세부견종'][i] = ', '.join(sub_breeds_list)
+            seoul_api_data.loc[i, '종'] = '믹스'
+            seoul_api_data.loc[i, '세부견종'] = ', '.join(sub_breeds_list)
         else:
-            seoul_api_data['종'][i] = '믹스'
-            seoul_api_data['세부견종'][i] = sub_breeds
+            seoul_api_data.loc[i, '종'] = '믹스'
+            seoul_api_data.loc[i, '세부견종'] = sub_breeds
     # 믹스 키워드가 있는 경우
     elif index_mix != -1:
-        seoul_api_data['종'][i] = '믹스'
-        if seoul_api_data['종'][i] == '믹스':
-            seoul_api_data['세부견종'][i] = '믹스'
+        seoul_api_data.loc[i, '종'] = '믹스'
+        if seoul_api_data.loc[i, '종'] == '믹스':
+            seoul_api_data.loc[i, '세부견종'] = '믹스'
         else:
-            seoul_api_data['세부견종'][i] = sample[:index_mix]
+            seoul_api_data.loc[i, '세부견종'] = sample[:index_mix]
     # 괄호와 믹스 키워드가 없는 경우
     else:
-        seoul_api_data['세부견종'][i] = sample
+        seoul_api_data.loc[i, '세부견종'] = sample
+
+
+# '특징' 컬럼 내에 각종 HTML 태그 제거 및 연속된 공백 제거
+seoul_api_data['특징'] = seoul_api_data['특징'].apply(
+    lambda x: re.sub(r'<.*?>|&[a-z]+;|\uFEFF|\u200B', '', x)  # HTML 태그, 엔티티, ZWNBSP 제거
+).apply(
+    lambda x: re.sub(r'\s+', ' ', x)  # 연속된 공백을 단일 공백으로 대체
+).str.strip()  # 앞뒤 공백 제거
 
 
 # 특정 키워드를 기준으로 텍스트를 정리
@@ -92,9 +161,9 @@ for i in range(len(seoul_api_data)):
 
     # 두 가지 케이스에 대한 텍스트 분리
     if index_wait != -1:
-      seoul_api_data['특징'][i] = sample[:index_wait+6]
+        seoul_api_data.loc[i, '특징'] = sample[:index_wait + 6]
     elif index_marketing != -1:
-      seoul_api_data['특징'][i] = sample[:index_marketing]
+        seoul_api_data.loc[i, '특징'] = sample[:index_marketing]
 
 
 # '특징'에서 색상 정보를 추출하여 새로운 '색상' 컬럼에 저장
@@ -103,15 +172,15 @@ for i in range(len(seoul_api_data)):
 
     colors = []
     if sample.find('흰') != -1 or sample.find('하얀') != -1 or sample.find('화이트') != -1:
-        colors.append('white')
+        colors.append('White')
     if sample.find('검정') != -1 or sample.find('검은') != -1  or sample.find('깜장') != -1 or sample.find('까만') != -1:
-        colors.append('black')
+        colors.append('Black')
     if sample.find('갈색') != -1 or sample.find('초코') != -1:
-        colors.append('brown')
+        colors.append('Brown')
     if sample.find('베이지') != -1 or sample.find('노랑') != -1 or sample.find('노란') != -1:
-        colors.append('beige')
+        colors.append('Beige')
     if sample.find('아이보리') != -1:
-        colors.append('ivory')
+        colors.append('Ivory')
 
     # 추출된 색상 정보를 ', '로 연결
     color_string = ', '.join(colors)
@@ -131,19 +200,23 @@ for i in range(len(seoul_api_data)):
         seoul_api_data.at[i, '색상'] = 'White'
 
 
-# '입양여부' 컬럼 추가 : 서울동물복지지원센터 입양대기동물 현황의 api 데이터이므로 모두 N로 처리
-seoul_api_data['입양여부'] = 'N'
+# '입양여부' 컬럼 정보 수정 : 서울동물복지지원센터 입양대기동물 현황의 api 데이터이므로 모두 N로 처리
+seoul_api_data['입양상태'] = 'N'
+
+
+# '중성화여부' 컬럼 추가 : 데이터 자체에서 중성화 여부를 알 수 없으므로 Unknown 처리
+seoul_api_data['중성화여부'] = 'U'
 
 
 # 불필요한 컬럼 삭제
 seoul_api_data = seoul_api_data.drop(columns=['ID'])
 
+
 # 일부 컬럼명 변경
 seoul_api_data = seoul_api_data.rename(columns={
     '종': '견종',
-    '태어난년도': '출생연도',
-    '크기': '체중',
-    '사진': '썸네일'
+    '무게': '체중',
+    '입양상태': '입양여부'
 })
